refactor: extract hook logic from BoaedList component
